refactor(runtime): persist submission handles
